API test updates in progress.
