+++ bin/api-tests
@@ -1,14 +1,35 @@
 #!/usr/bin/env bash
 
-
-npm run package-test
+# npm run package-test
 npm run package-api-test
 
-# Run docker-compose with development configuration
-docker container rm docker_postgresql_1 --force
+# Remove prisma/lpwanserver containers and volumes
+docker container rm dev_lpwanserver_prisma --force
+docker container rm dev_lpwanserver_postgres --force
+docker volume rm dev_lpwanserver_postgres --force
 
+# Remove LoRa Server postgres container and volume
+docker container rm dev_lpwanserver_loraserver_postgres --force
+docker volume rm docker_lpwanserver_loraserver_postgres --force
+
+# Start Prisma and LoRa Server
 docker-compose \
-  -f ./docker/docker-compose.loraserver.yml \
+  -f ./prisma/dev/docker-compose.yaml up -d
+
+sleep 2s
+
+# Deploy Prisma datamodel
+cd prisma/dev && prisma deploy && cd ../..
+
+# Start Lora Server and run tests
+docker-compose \
   -f ./docker/docker-compose.api-test.yml up \
   --abort-on-container-exit \
   --exit-code-from api-test
+
+# Stop Prisma and LoRa Server
+docker-compose \
+  -f ./prisma/dev/docker-compose.yaml \
+  -f ./docker/docker-compose.loraserver.yml down
+
+  # -f ./docker/docker-compose.loraserver.yml \
