+++ docker/docker-compose.api-test.yml
@@ -1,21 +1,27 @@
-version: '3'
+version: '3.5'
 
 services:
   api-test:
-    image: lpwanserver/api-test
-    environment:
-      - NODE_ENV=e2e
-      - PORT=3200
-    depends_on:
-      - loraserver1
-      - lora_appserver1
+    image: lpwanserver/api-test:latest
+    networks:
+      - lpwanserver_net
     ports:
       - '3200:3200'
-
-  postgresql:
-    volumes:
-      - ./configuration/postgresql/initdb-demo:/docker-entrypoint-initdb.d
-
-  redis:
-    volumes:
-      - ../data/demo_data/redis:/data
+    environment:
+      - NODE_ENV=e2e
+      - port=3200
+      - prisma_protocol=http
+      - prisma_host=prisma
+      - prisma_port=4466
+    # depends_on:
+    #   - loraserver1
+    #   - lora_appserver1
+  # postgresql:
+  #   volumes:
+  #     - ./configuration/postgresql/initdb-demo:/docker-entrypoint-initdb.d
+  # redis:
+  #   volumes:
+  #     - ../data/demo_data/redis:/data
+networks:
+  lpwanserver_net:
+    name: lpwanserver_net
