+++ docker/docker-compose.loraserver.yml
@@ -61,11 +61,11 @@ services:
       - lpwanserver_loraserver_postgres:/var/lib/postgresql/data
 
   redis:
+    image: redis:4-alpine
     networks:
       - lpwanserver_net
     ports:
       - 6379
-    image: redis:4-alpine
 
   mosquitto:
     image: eclipse-mosquitto
