+++ rest/models/INetworkProtocol.js
@@ -1,6 +1,7 @@
 // Configuration access.
 var nconf = require('nconf')
 const appLogger = require('../lib/appLogger')
+const path = require('path')
 
 //* *****************************************************************************
 // The NetworkProtocol interface.
@@ -12,16 +13,15 @@ function NetworkProtocol () {
 }
 
 NetworkProtocol.prototype.retrieveNetworkProtocols = async function (options) {
-  let recs = await this.impl.retrieveNetworkProtocols(options)
-  appLogger.log(recs)
-  let len = recs.records.length
-  for (let i = 0; i < len; i++) {
-    let rec = recs.records[i]
-    let handler = require('../networkProtocols/' + rec.protocolHandler)
-    rec.metaData = handler.metaData
-    recs.records[i] = rec
+  let result = await this.impl.retrieveNetworkProtocols(options)
+  const handlerDir = path.join(__dirname, '../networkProtocols/handlers')
+  return {
+    ...result,
+    records: result.records.map(x => {
+      const { metaData } = require(path.join(handlerDir, x.protocolHandler))
+      return { ...x, metaData }
+    })
   }
-  return recs
 }
 
 NetworkProtocol.prototype.retrieveNetworkProtocol = async function (id) {
