+++ rest/models/IUser.js
@@ -19,11 +19,12 @@ function User () {
   this.reverseRoles = reverseRoles
 
   // Load the roles
-  let roleList = this.impl.getRoles()
-  for (var i = 0; i < roleList.length; ++i) {
-    roles[ roleList[ i ].name ] = roleList[ i ].id
-    reverseRoles[ roleList[ i ].id ] = roleList[ i ].name
-  }
+  this.impl.getRoles().then(roleList => {
+    for (var i = 0; i < roleList.length; ++i) {
+      roles[ roleList[ i ].name ] = roleList[ i ].id
+      reverseRoles[ roleList[ i ].id ] = roleList[ i ].name
+    }
+  })
 }
 
 User.prototype.authorizeUser = function (username, password) {
