+++ rest/networkProtocols/handlers/LoRaOpenSource_1.js
@@ -2,7 +2,8 @@ var request = require('request')
 var nconf = require('nconf')
 
 // General libraries in use in this module.
-var appLogger = require('../lib/appLogger.js')
+var appLogger = require('../../lib/appLogger.js')
+// var appLogger = appLogger.logger
 
 /**
  * The LoRa Server Protocol Handler Module
@@ -15,11 +16,11 @@ module.exports = {}
 module.exports.activeApplicationNetworkProtocols = {}
 module.exports.metaData =
   {
-    protocolHandlerName: 'LoRa Server 2.0',
+    protocolHandlerName: 'LoRa Server',
     version:
       {
-        versionText: 'Version 2.0',
-        versionValue: '2.0'
+        versionText: 'Version 1.0',
+        versionValue: '1.0'
       },
     networkType: 'Lora',
     oauthUrl: '',
@@ -49,39 +50,13 @@ module.exports.metaData =
     ]
   }
 
-module.exports.register = async function (networkProtocols) {
-  appLogger.log('LoraOpenSource:register', 'warn')
-  return new Promise(async function (resolve, reject) {
-    let me = {
-      name: 'LoRa Server',
-      networkTypeId: 1,
-      protocolHandler: 'LoRaOpenSource_2.js',
-      networkProtocolVersion: '2.0'
-    }
-    setTimeout(function () {
-      networkProtocols.retrieveNetworkProtocols({ search: me.name, networkProtocolVersion: '1.0' })
-        .then(master => {
-          console.log('Found Master')
-          me.masterProtocol = master.records[0].id
-          networkProtocols.upsertNetworkProtocol(me)
-            .then(() => {
-              resolve()
-            })
-            .catch(err => {
-              reject(err)
-            })
-        })
-        .catch(err => {
-          // No master
-          networkProtocols.upsertNetworkProtocol(me)
-            .then(() => {
-              resolve()
-            })
-            .catch(err => {
-              reject(err)
-            })
-        })
-    }, 1000)
+module.exports.register = async function register (networkProtocols) {
+  appLogger.log('LoraOpenSource:register')
+  await networkProtocols.upsertNetworkProtocol({
+    name: 'LoRa Server',
+    networkTypeId: 1,
+    protocolHandler: 'LoRaOpenSource_1.js',
+    networkProtocolVersion: '1.0'
   })
 }
 
@@ -125,11 +100,10 @@ module.exports.test = function (network, loginData) {
           }
         }
         else {
-          appLogger.log('Test Error: ', 'error')
+          appLogger.log('Test Error: ' + error)
           if (response && response.statusCode) {
             appLogger.log(response.statusCode)
           }
-          appLogger.log(error, 'error')
           reject(error)
         }
       })
@@ -295,10 +269,6 @@ async function getCompanyAccount (dataAPI, network, companyId, generateIfMissing
   return secData
 };
 
-function makeNetworkDataKey (networkId, dataName) {
-  return 'nwk:' + networkId + '/' + dataName
-}
-
 function makeCompanyDataKey (companyId, dataName) {
   return 'co:' + companyId + '/' + dataName
 }
@@ -338,8 +308,7 @@ function getANetworkServerID (network, connection) {
     request(options, async function (error, response, body) {
       if (error || response.statusCode >= 400) {
         if (error) {
-          appLogger.log('Error on get Network Server: ', 'error')
-          appLogger.log(error, 'error')
+          appLogger.log('Error on get Network Server: ' + error)
           reject(error)
         }
         else {
@@ -387,8 +356,7 @@ function getNetworkServerById (network, networkServerId, connection, dataAPI) {
     request(options, async function (error, response, body) {
       if (error || response.statusCode >= 400) {
         if (error) {
-          appLogger.log('Error on get Network Server: ', 'error')
-          appLogger.log(error, 'error')
+          appLogger.log('Error on get Network Server: ' + error)
           reject(error)
         }
         else {
@@ -401,14 +369,9 @@ function getNetworkServerById (network, networkServerId, connection, dataAPI) {
         }
       }
       else {
-        let res = ''
-        if (typeof body === 'object') {
-          res = body
-        }
-        else {
-          res = JSON.parse(body)
-        }
-        appLogger.log(res, 'warn')
+        // Convert text to JSON array, use id from first element
+        var res = JSON.parse(body)
+        appLogger.log(res)
         resolve(res)
       }
     })
@@ -430,18 +393,17 @@ function getDeviceProfileById (network, dpId, connection, dataAPI) {
       'secureProtocol': 'TLSv1_2_method',
       'rejectUnauthorized': false
     }
-    appLogger.log(options, 'warn')
+    appLogger.log(options)
     request(options, async function (error, response, body) {
       if (error || response.statusCode >= 400) {
         if (error) {
-          appLogger.log('Error on get Device Profile: ', 'error')
-          appLogger.log(error, 'error')
+          appLogger.log('Error on get Device Profile: ' + error)
           reject(error)
         }
         else {
           var bodyObj = JSON.parse(response.body)
           appLogger.log(bodyObj)
-          appLogger.log(options, 'warn')
+          appLogger.log(options)
           appLogger.log('Error on get Device Profile: ' +
             bodyObj.error +
             ' (' + response.statusCode + ')')
@@ -473,12 +435,11 @@ function getRemoteDeviceById (network, deviceId, connection) {
       'secureProtocol': 'TLSv1_2_method',
       'rejectUnauthorized': false
     }
-    appLogger.log(options, 'warn')
+    appLogger.log(options)
     request(options, async function (error, response, body) {
       if (error || response.statusCode >= 400) {
         if (error) {
-          appLogger.log('Error on get Device: ', 'error')
-          appLogger.log(error, 'error')
+          appLogger.log('Error on get Device: ' + error)
           reject(error)
         }
         else {
@@ -513,12 +474,11 @@ function getRemoteDeviceKey (network, device, connection) {
       'secureProtocol': 'TLSv1_2_method',
       'rejectUnauthorized': false
     }
-    appLogger.log(options, 'warn')
+    appLogger.log(options)
     request(options, async function (error, response, body) {
       if (error || response.statusCode >= 400) {
         if (error) {
-          appLogger.log('Error on get Device Keys: ', 'error')
-          appLogger.log(error, 'error')
+          appLogger.log('Error on get Device Keys: ' + error)
           reject(error)
         }
         else if (response.statusCode === 404) {
@@ -548,7 +508,7 @@ function getRemoteDeviceActivation (network, device, connection) {
   return new Promise(async function (resolve, reject) {
     let options = {}
     options.method = 'GET'
-    options.url = network.baseUrl + '/devices/' + device.device.devEUI + '/activation'
+    options.url = network.baseUrl + '/devices/' + device.devEUI + '/activation'
     options.headers = {
       'Content-Type': 'application/json',
       'Authorization': 'Bearer ' + connection
@@ -557,7 +517,7 @@ function getRemoteDeviceActivation (network, device, connection) {
       'secureProtocol': 'TLSv1_2_method',
       'rejectUnauthorized': false
     }
-    appLogger.log(options, 'warn')
+    appLogger.log(options)
     request(options, async function (error, response, body) {
       if (error || response.statusCode >= 400) {
         if (error) {
@@ -580,7 +540,8 @@ function getRemoteDeviceActivation (network, device, connection) {
       }
       else {
         let keys = JSON.parse(body)
-        device.device.deviceActivation = keys.deviceActivation
+        appLogger.log(keys, 'info')
+        device.deviceActivation = keys
         resolve(device)
       }
     })
@@ -607,8 +568,7 @@ function getServiceProfileById (network, serviceProfileId, connection, dataAPI) 
     request(options, async function (error, response, body) {
       if (error || response.statusCode >= 400) {
         if (error) {
-          appLogger.log('Error on get Service Profile: ', 'error')
-          appLogger.log(error, 'error')
+          appLogger.log('Error on get Service Profile: ' + error)
           reject(error)
         }
         else {
@@ -646,12 +606,11 @@ function getApplicationById (network, applicationId, connection) {
       'rejectUnauthorized': false
     }
 
-    appLogger.log(options, 'warn')
+    appLogger.log(options)
     request(options, async function (error, response, body) {
       if (error || response.statusCode >= 400) {
         if (error) {
-          appLogger.log('Error on get Application: ', 'error')
-          appLogger.log(error, 'error')
+          appLogger.log('Error on get Application: ' + error)
           reject(error)
         }
         else {
@@ -675,7 +634,7 @@ function getApplicationById (network, applicationId, connection) {
 
 // Get the Service Profile a for a Remote Org.
 function getServiceProfileForOrg (network, orgId, companyId, connection, dataAPI) {
-  appLogger.log('LoRaOpenSource: getServiceProfileForOrg', 'debug')
+  appLogger.log('LoRaOpenSource: getNetworkServerForRemoteOrganization')
   return new Promise(async function (resolve, reject) {
     var spOptions = {}
     spOptions.method = 'GET'
@@ -693,8 +652,7 @@ function getServiceProfileForOrg (network, orgId, companyId, connection, dataAPI
     request(spOptions, async function (error, response, body) {
       if (error || response.statusCode >= 400) {
         if (error) {
-          appLogger.log('Error getting Service Profile: ', 'error')
-          appLogger.log(error, 'error')
+          appLogger.log('Error getting Service Profile: ' + error)
           reject(error)
         }
         else {
@@ -714,12 +672,11 @@ function getServiceProfileForOrg (network, orgId, companyId, connection, dataAPI
         }
         else {
           let serviceProfile = body.result[0]
-          appLogger.log(serviceProfile, 'warn')
           await dataAPI.putProtocolDataForKey(
             network.id,
             network.networkProtocol.id,
             makeCompanyDataKey(companyId, 'coSPId'),
-            serviceProfile.id)
+            serviceProfile.serviceProfileID)
           // TODO: Remove this HACK.  Should not store the service profile locally
           //       in case it gets changed on the remote server.
           await dataAPI.putProtocolDataForKey(
@@ -836,8 +793,7 @@ module.exports.connect = function (network, loginData) {
     options.agentOptions = { 'secureProtocol': 'TLSv1_2_method', 'rejectUnauthorized': false }
     request(options, function (error, response, body) {
       if (error) {
-        appLogger.log('Error on signin: ', 'error')
-        appLogger.log(error, 'error')
+        appLogger.log('Error on signin: ' + error)
         reject(error)
       }
       else if (response.statusCode >= 400 || response.statusCode === 301) {
@@ -912,11 +868,9 @@ module.exports.addCompany = function (sessionData, network, companyId, dataAPI) 
     // No need to allow gateways on organizations we create (may be changed
     // locally by network admin, and that's fine - just start w/o gateways).
     options.json = {
-      organization: {
-        'name': company.name,
-        'displayName': company.name,
-        'canHaveGateways': false
-      }
+      'name': company.name,
+      'displayName': company.name,
+      'canHaveGateways': false
     }
     options.agentOptions = {
       'secureProtocol': 'TLSv1_2_method',
@@ -926,8 +880,7 @@ module.exports.addCompany = function (sessionData, network, companyId, dataAPI) 
     request(options, async function (error, response, body) {
       if (error || response.statusCode >= 400) {
         if (error) {
-          appLogger.log('Error on create company ' + company.name + ': ', 'error')
-          appLogger.log(error, 'error')
+          appLogger.log('Error on create company ' + company.name + ': ' + error)
           reject(error)
         }
         else {
@@ -940,7 +893,6 @@ module.exports.addCompany = function (sessionData, network, companyId, dataAPI) 
       else {
         try {
           // Save the company ID from the remote network.
-          appLogger.log(body)
           await dataAPI.putProtocolDataForKey(network.id,
             network.networkProtocol.id,
             makeCompanyDataKey(company.id, 'coNwkId'),
@@ -960,21 +912,19 @@ module.exports.addCompany = function (sessionData, network, companyId, dataAPI) 
           var creds = await getCompanyAccount(dataAPI, network, companyId, true)
 
           userOptions.json = {
+            'username': creds.username,
             'password': creds.password,
-            user: {
-              'username': creds.username,
-              'isActive': true,
-              'isAdmin': false,
-              'sessionTTL': 0,
-              'email': 'fake@emailaddress.com',
-              'note': 'Created by and for LPWAN Server'
-            },
+            'isActive': true,
+            'isAdmin': false,
+            'sessionTTL': 0,
             'organizations': [
               {
                 'isAdmin': true,
                 'organizationID': networkCoId
               }
-            ]
+            ],
+            'email': 'fake@emailaddress.com',
+            'note': 'Created by and for LPWAN Server'
           }
           userOptions.agentOptions = {
             'secureProtocol': 'TLSv1_2_method',
@@ -986,8 +936,7 @@ module.exports.addCompany = function (sessionData, network, companyId, dataAPI) 
               if (error) {
                 appLogger.log('Error creating ' + company.name +
                   "'s admin user " + userOptions.json.username +
-                  ': ', 'error')
-                appLogger.log(error, 'error')
+                  ': ' + error)
                 reject(error)
                 return
               }
@@ -1021,10 +970,10 @@ module.exports.addCompany = function (sessionData, network, companyId, dataAPI) 
             }
 
             spOptions.json = {
+              'name': 'defaultForLPWANServer',
+              'networkServerID': networkServerId,
+              'organizationID': networkCoId,
               'serviceProfile': {
-                'name': 'defaultForLPWANServer',
-                'networkServerID': networkServerId,
-                'organizationID': networkCoId,
                 'addGWMetadata': true,
                 'devStatusReqFreq': 1,
                 'dlBucketSize': 0,
@@ -1045,8 +994,7 @@ module.exports.addCompany = function (sessionData, network, companyId, dataAPI) 
             request(spOptions, async function (error, response, body) {
               if (error || response.statusCode >= 400) {
                 if (error) {
-                  appLogger.log('Error creating default Service Profile: ', 'error')
-                  appLogger.log(error, 'error')
+                  appLogger.log('Error creating default Service Profile: ' + error)
                   reject(error)
                 }
                 else {
@@ -1062,7 +1010,7 @@ module.exports.addCompany = function (sessionData, network, companyId, dataAPI) 
                   network.id,
                   network.networkProtocol.id,
                   makeCompanyDataKey(company.id, 'coSPId'),
-                  body.id)
+                  body.serviceProfileID)
                 // TODO: Remove this HACK.  Should not store the service profile locally
                 //       in case it gets changed on the remote server.
                 await dataAPI.putProtocolDataForKey(
@@ -1129,8 +1077,7 @@ module.exports.getCompany = function (sessionData, network, companyId, dataAPI) 
 
     request(options, function (error, response, body) {
       if (error) {
-        appLogger.log('Error getting company ' + company.name + ': ', 'error')
-        appLogger.log(error, 'error')
+        appLogger.log('Error getting company ' + company.name + ': ' + error)
         reject(error)
       }
       else {
@@ -1191,8 +1138,7 @@ module.exports.updateCompany = function (sessionData, network, companyId, dataAP
     request(options, function (error, response, body) {
       if (error) {
         appLogger.log('Error updating company ' + company.name +
-          ': ', 'error')
-        appLogger.log(error, 'error')
+          ': ' + error)
         reject(error)
       }
       else {
@@ -1251,8 +1197,7 @@ module.exports.deleteCompany = function (sessionData, network, companyId, dataAP
     }
     request(userOptions, function (error, response, body) {
       if (error) {
-        appLogger.log("Error on delete company's admin user: ", 'error')
-        appLogger.log(error, 'error')
+        appLogger.log("Error on delete company's admin user: " + error)
         reject(error)
       }
       else {
@@ -1271,8 +1216,7 @@ module.exports.deleteCompany = function (sessionData, network, companyId, dataAP
 
         request(options, async function (error, response, body) {
           if (error) {
-            appLogger.log('Error on delete company: ', 'error')
-            appLogger.log(error, 'error')
+            appLogger.log('Error on delete company: ' + error)
             reject(error)
           }
           else {
@@ -1435,13 +1379,11 @@ module.exports.pushDeviceProfile = function (sessionData, network, deviceProfile
   let me = this
   return new Promise(async function (resolve, reject) {
     // See if it already exists
-    appLogger.log(deviceProfile, 'info')
     dataAPI.getProtocolDataForKey(
       network.id,
       network.networkProtocol.id,
       makeDeviceProfileDataKey(deviceProfile.id, 'dpNwkId'))
       .then(dpNetworkId => {
-        appLogger.log(dpNetworkId, 'info')
         if (isUpdate && dpNetworkId) {
           me.updateDeviceProfile(sessionData, network, deviceProfile.id, dataAPI)
             .then(resolve)
@@ -1458,8 +1400,6 @@ module.exports.pushDeviceProfile = function (sessionData, network, deviceProfile
           })
         }
         else {
-          appLogger.log(dpNetworkId + '')
-
           reject(new Error('Something bad happened with the Protocol Table'))
         }
       })
@@ -1502,6 +1442,7 @@ module.exports.pushDevices = function (sessionData, network, modelAPI, dataAPI) 
 
 module.exports.pushDevice = function (sessionData, network, device, dataAPI, update = true) {
   let me = this
+  appLogger.log('Hello?')
   return new Promise(async function (resolve, reject) {
     // See if it already exists
     dataAPI.getProtocolDataForKey(
@@ -1619,18 +1560,17 @@ module.exports.setupOrganization = function (sessionData, network, modelAPI, dat
       'rejectUnauthorized': false
     }
 
-    appLogger.log(options, 'warn')
+    appLogger.log(options)
     request(options, function (error, response, body) {
       if (error) {
-        appLogger.log('Error getting operator ' + company.name + ' from network ' + network.name + ': ', 'error')
-        appLogger.log(error, 'error')
+        appLogger.log('Error getting operator ' + company.name + ' from network ' + network.name + ': ' + error)
         reject(error)
       }
       else {
         body = JSON.parse(body)
         appLogger.log(body)
-        if (body.totalCount === '0') {
-          appLogger.log('Adding company to network ', 'warn')
+        if (body.totalCount === 0) {
+          appLogger.log('Adding company')
           me.addCompany(sessionData, network, company.id, dataAPI)
             .then((networkSettings) => {
               appLogger.log(networkSettings)
@@ -1658,15 +1598,15 @@ module.exports.setupOrganization = function (sessionData, network, modelAPI, dat
               getNetworkServerById(network, networkSettings.networkServerID, sessionData.connection, dataAPI)
                 .then(networkServer => {
                   appLogger.log(networkServer, 'warn')
-                  lora1NetworkSettings.serviceProfileId = networkSettings.id
+                  lora1NetworkSettings.serviceProfileId = networkSettings.serviceProfileID
                   lora1NetworkSettings.networkServerId = networkSettings.networkServerID
                   lora1NetworkSettings.organizationId = organization.id
                   lora1NetworkSettings.networkId = network.id
 
                   companyNtl.networkSettings.serviceProfile = { region: networkServer.region }
                   companyNtl.networkSettings[network.name] = lora1NetworkSettings
-                  appLogger.log(companyNtl, 'warn')
 
+                  appLogger.log(companyNtl, 'warn')
                   modelAPI.companyNetworkTypeLinks.updateRemoteCompanyNetworkTypeLink(companyNtl)
                     .then((result) => {
                       appLogger.log(result)
@@ -1717,11 +1657,10 @@ module.exports.pullDeviceProfiles = function (sessionData, network, modelAPI, co
       'rejectUnauthorized': false
     }
 
-    appLogger.log(options, 'warn')
+    appLogger.log(options)
     request(options, function (error, response, body) {
       if (error) {
-        appLogger.log('Error pulling device profiles from network ' + network.name + ': ', 'error')
-        appLogger.log(error, 'error')
+        appLogger.log('Error pulling device profiles from network ' + network.name + ': ' + error)
         reject(error)
       }
       else {
@@ -1736,7 +1675,6 @@ module.exports.pullDeviceProfiles = function (sessionData, network, modelAPI, co
         }
         Promise.all(promiseList)
           .then((devices) => {
-            appLogger.log(devices, 'info')
             resolve(devices)
           })
           .catch(err => {
@@ -1750,11 +1688,11 @@ module.exports.pullDeviceProfiles = function (sessionData, network, modelAPI, co
 
 module.exports.addRemoteDeviceProfile = function (sessionData, limitedRemoteDeviceProfile, network, modelAPI, dataAPI) {
   return new Promise(async function (resolve, reject) {
-    getDeviceProfileById(network, limitedRemoteDeviceProfile.id, sessionData.connection)
+    getDeviceProfileById(network, limitedRemoteDeviceProfile.deviceProfileID, sessionData.connection)
       .then((remoteDeviceProfile) => {
-        appLogger.log('Adding ' + remoteDeviceProfile.deviceProfile.name)
+        appLogger.log('Adding ' + remoteDeviceProfile.name)
         appLogger.log(remoteDeviceProfile)
-        modelAPI.deviceProfiles.retrieveDeviceProfiles({ search: remoteDeviceProfile.deviceProfile.name })
+        modelAPI.deviceProfiles.retrieveDeviceProfiles({ search: remoteDeviceProfile.name })
           .then(existingDeviceProfile => {
             if (existingDeviceProfile.totalCount > 0) {
               existingDeviceProfile = existingDeviceProfile.records[0]
@@ -1763,19 +1701,19 @@ module.exports.addRemoteDeviceProfile = function (sessionData, limitedRemoteDevi
               dataAPI.putProtocolDataForKey(network.id,
                 network.networkProtocol.id,
                 makeDeviceProfileDataKey(existingDeviceProfile.id, 'dpNwkId'),
-                remoteDeviceProfile.deviceProfile.id)
+                remoteDeviceProfile.deviceProfile.deviceProfileID)
 
               resolve({
                 localDeviceProfile: existingDeviceProfile.id,
-                remoteDeviceProfile: remoteDeviceProfile.deviceProfile.id
+                remoteDeviceProfile: remoteDeviceProfile.deviceProfile.deviceProfileID
               })
             }
             else {
-              appLogger.log('creating ' + remoteDeviceProfile.deviceProfile.name)
+              appLogger.log('creating ' + remoteDeviceProfile.name)
               let networkSpecificDeviceProfileInformation = normalizeDeviceProfileData(remoteDeviceProfile)
               appLogger.log(networkSpecificDeviceProfileInformation, 'error')
               modelAPI.deviceProfiles.createRemoteDeviceProfile(network.networkType.id, 2,
-                remoteDeviceProfile.deviceProfile.name, 'Device Profile managed by LPWAN Server, perform changes via LPWAN',
+                remoteDeviceProfile.name, 'Device Profile managed by LPWAN Server, perform changes via LPWAN',
                 networkSpecificDeviceProfileInformation)
                 .then((existingDeviceProfile) => {
                   appLogger.log(existingDeviceProfile)
@@ -1783,29 +1721,26 @@ module.exports.addRemoteDeviceProfile = function (sessionData, limitedRemoteDevi
                   dataAPI.putProtocolDataForKey(network.id,
                     network.networkProtocol.id,
                     makeDeviceProfileDataKey(existingDeviceProfile.id, 'dpNwkId'),
-                    remoteDeviceProfile.deviceProfile.id)
+                    remoteDeviceProfile.deviceProfile.deviceProfileID)
 
                   resolve({
                     localDeviceProfile: existingDeviceProfile.id,
-                    remoteDeviceProfile: remoteDeviceProfile.deviceProfile.id
+                    remoteDeviceProfile: remoteDeviceProfile.deviceProfile.deviceProfileID
                   })
                 })
                 .catch((error) => {
                   appLogger.log(error)
-                  appLogger.log(error, 'error')
                   reject(error)
                 })
             }
           })
           .catch((error) => {
             appLogger.log(error)
-            appLogger.log(error, 'error')
             reject(error)
           })
       })
       .catch((error) => {
         appLogger.log(error)
-        appLogger.log(error, 'error')
         reject(error)
       })
   })
@@ -1834,11 +1769,10 @@ module.exports.pullApplications = function (sessionData, network, modelAPI, data
       'rejectUnauthorized': false
     }
 
-    appLogger.log(options, 'warn')
+    appLogger.log(options)
     request(options, function (error, response, body) {
       if (error) {
-        appLogger.log('Error pulling applications from network ' + network.name + ': ', 'error')
-        appLogger.log(error, 'error')
+        appLogger.log('Error pulling applications from network ' + network.name + ': ' + error)
         reject(error)
       }
       else {
@@ -1867,14 +1801,14 @@ function addRemoteApplication (sessionData, limitedRemoteApplication, network, m
   let me = this
   return new Promise(async function (resolve, reject) {
     let remoteApplication = await getApplicationById(network, limitedRemoteApplication.id, sessionData.connection)
-    appLogger.log(remoteApplication, 'error')
-    let existingApplication = await modelAPI.applications.retrieveApplications({ search: remoteApplication.application.name })
+    appLogger.log(remoteApplication)
+    let existingApplication = await modelAPI.applications.retrieveApplications({ search: remoteApplication.name })
     if (existingApplication.totalCount > 0) {
       existingApplication = existingApplication.records[0]
       appLogger.log(existingApplication.name + ' already exists')
     }
     else {
-      existingApplication = await modelAPI.applications.createApplication(remoteApplication.application.name, remoteApplication.application.description, 2, 1, 'http://set.me.to.your.real.url:8888')
+      existingApplication = await modelAPI.applications.createApplication(remoteApplication.name, remoteApplication.description, 2, 1, 'http://set.me.to.your.real.url:8888')
       appLogger.log('Created ' + existingApplication.name)
     }
 
@@ -1890,9 +1824,9 @@ function addRemoteApplication (sessionData, limitedRemoteApplication, network, m
       await dataAPI.putProtocolDataForKey(network.id,
         network.networkProtocol.id,
         makeApplicationDataKey(existingApplication.id, 'appNwkId'),
-        remoteApplication.application.id)
+        remoteApplication.id)
     }
-    resolve({ localApplication: existingApplication.id, remoteApplication: remoteApplication.application.id })
+    resolve({ localApplication: existingApplication.id, remoteApplication: remoteApplication.id })
   })
 }
 
@@ -1909,11 +1843,10 @@ function addRemoteApplication (sessionData, limitedRemoteApplication, network, m
  * @returns {Promise<any>}
  */
 module.exports.pullDevices = function (sessionData, network, remoteApplicationId, localApplicationId, dpMap, modelAPI, dataAPI) {
-  let me = this
   return new Promise(async function (resolve, reject) {
     let options = {}
     options.method = 'GET'
-    options.url = network.baseUrl + '/devices' + '?limit=9999&offset=0&applicationID=' + remoteApplicationId
+    options.url = network.baseUrl + '/applications/' + remoteApplicationId + '/devices' + '?limit=9999&offset=0'
     options.headers = {
       'Content-Type': 'application/json',
       'Authorization': 'Bearer ' + sessionData.connection
@@ -1923,11 +1856,10 @@ module.exports.pullDevices = function (sessionData, network, remoteApplicationId
       'rejectUnauthorized': false
     }
 
-    appLogger.log(options, 'error')
+    appLogger.log(options)
     request(options, function (error, response, body) {
       if (error) {
-        appLogger.log('Error pulling devices from network ' + network.name + ': ', 'error')
-        appLogger.log(error, 'error')
+        appLogger.log('Error pulling devices from network ' + network.name + ': ' + error)
         reject(error)
       }
       else {
@@ -1937,7 +1869,7 @@ module.exports.pullDevices = function (sessionData, network, remoteApplicationId
         let promiseList = []
         for (let index in devices) {
           let device = devices[index]
-          promiseList.push(me.addRemoteDevice(sessionData, device, network, localApplicationId, dpMap, modelAPI, dataAPI))
+          promiseList.push(addRemoteDevice(sessionData, device, network, localApplicationId, dpMap, modelAPI, dataAPI))
         }
         Promise.all(promiseList)
           .then((apps) => {
@@ -1952,12 +1884,12 @@ module.exports.pullDevices = function (sessionData, network, remoteApplicationId
   })
 }
 
-module.exports.addRemoteDevice = function (sessionData, limitedRemoteDevice, network, applicationId, dpMap, modelAPI, dataAPI) {
+function addRemoteDevice (sessionData, limitedRemoteDevice, network, applicationId, dpMap, modelAPI, dataAPI) {
   return new Promise(async function (resolve, reject) {
     let remoteDevice = await getRemoteDeviceById(network, limitedRemoteDevice.devEUI, sessionData.connection, dataAPI)
-    appLogger.log('Adding ' + remoteDevice.device.name)
+    appLogger.log('Adding ' + remoteDevice.name)
     appLogger.log(remoteDevice)
-    let deviceProfileId = dpMap.find(o => o.remoteDeviceProfile === remoteDevice.device.deviceProfileID)
+    let deviceProfileId = dpMap.find(o => o.remoteDeviceProfile === remoteDevice.deviceProfileID)
     let deviceProfile = await dataAPI.getDeviceProfileById(deviceProfileId.localDeviceProfile)
     if (deviceProfile.supportsJoin) {
       remoteDevice = await getRemoteDeviceKey(network, remoteDevice, sessionData.connection)
@@ -1966,15 +1898,15 @@ module.exports.addRemoteDevice = function (sessionData, limitedRemoteDevice, net
       remoteDevice = await getRemoteDeviceActivation(network, remoteDevice, sessionData.connection)
     }
 
-    let existingDevice = await modelAPI.devices.retrieveDevices({ search: remoteDevice.device.name })
+    let existingDevice = await modelAPI.devices.retrieveDevices({ search: remoteDevice.name })
     appLogger.log(existingDevice)
     if (existingDevice.totalCount > 0) {
       existingDevice = existingDevice.records[0]
       appLogger.log(existingDevice.name + ' already exists')
     }
     else {
-      appLogger.log('creating ' + remoteDevice.device.name)
-      existingDevice = await modelAPI.devices.createDevice(remoteDevice.device.name, remoteDevice.device.description, applicationId)
+      appLogger.log('creating ' + remoteDevice.name)
+      existingDevice = await modelAPI.devices.createDevice(remoteDevice.name, remoteDevice.description, applicationId)
       appLogger.log('Created ' + existingDevice.name)
     }
 
@@ -1985,13 +1917,14 @@ module.exports.addRemoteDevice = function (sessionData, limitedRemoteDevice, net
     else {
       appLogger.log('creating Network Link for ' + existingDevice.name)
       let normalizedDeviceSettings = normalizeDeviceData(remoteDevice)
+      appLogger.log(normalizedDeviceSettings, 'info')
       existingDeviceNTL = await modelAPI.deviceNetworkTypeLinks.createRemoteDeviceNetworkTypeLink(existingDevice.id, network.networkType.id, deviceProfileId.localDeviceProfile, normalizedDeviceSettings, 2)
       appLogger.log(existingDeviceNTL)
     }
     dataAPI.putProtocolDataForKey(network.id,
       network.networkProtocol.id,
       makeDeviceDataKey(existingDevice.id, 'devNwkId'),
-      remoteDevice.device.devEUI)
+      remoteDevice.devEUI)
     resolve(existingDevice.id)
   })
 }
@@ -2020,17 +1953,15 @@ module.exports.pullIntegrations = function (sessionData, network, remoteApplicat
       'secureProtocol': 'TLSv1_2_method',
       'rejectUnauthorized': false
     }
-    appLogger.log(options, 'error')
+    appLogger.log(options)
     request(options, function (error, response, body) {
       if (error) {
-        appLogger.log('Error pulling integrations from network ' + network.name + ': ', 'error')
-        appLogger.log(error, 'error')
+        appLogger.log('Error pulling integrations from network ' + network.name + ': ' + error)
         reject(error)
       }
       else {
         let integration = JSON.parse(body)
-        integration = integration.integration
-        appLogger.log(integration, 'warn')
+        appLogger.log(integration)
         var deliveryURL = 'api/ingest/' + localApplicationId + '/' + network.id
         var reportingUrl = nconf.get('base_url') + deliveryURL
 
@@ -2040,12 +1971,12 @@ module.exports.pullIntegrations = function (sessionData, network, remoteApplicat
         else {
           modelAPI.applications.retrieveApplication(localApplicationId)
             .then(appToUpdate => {
-              appToUpdate.baseUrl = integration.uplinkDataURL
-              appLogger.log(appToUpdate, 'warn')
+              appToUpdate.baseUrl = integration.dataUpURL
+              appLogger.log(appToUpdate)
               delete appToUpdate.networks
               modelAPI.applications.updateApplication(appToUpdate)
                 .then(result => {
-                  appLogger.log(result, 'warn')
+                  appLogger.log(result)
                   let options2 = {}
                   options2.method = 'PUT'
                   options2.url = network.baseUrl + '/applications/' + remoteApplicationId + '/integrations/http'
@@ -2059,23 +1990,19 @@ module.exports.pullIntegrations = function (sessionData, network, remoteApplicat
                   }
 
                   options2.json = {
-                    integration: {
-                      ackNotificationURL: reportingUrl,
-                      uplinkDataURL: reportingUrl,
-                      errorNotificationURL: reportingUrl,
-                      id: integration.id,
-                      joinNotificationURL: reportingUrl
-                    }
+                    ackNotificationURL: reportingUrl,
+                    dataUpURL: reportingUrl,
+                    errorNotificationURL: reportingUrl,
+                    id: integration.id,
+                    joinNotificationURL: reportingUrl
                   }
                   appLogger.log(options2)
                   request(options2, function (error, response, body) {
                     if (error) {
-                      appLogger.log('Error updating integrations from network ' + network.name + ': ', 'error')
-                      appLogger.log(error, 'error')
+                      appLogger.log('Error updating integrations from network ' + network.name + ': ' + error)
                       reject(error)
                     }
                     else {
-                      appLogger.log('Success Setting Integrations', 'warn')
                       resolve()
                     }
                   })
@@ -2142,16 +2069,12 @@ module.exports.addApplication = function (sessionData, network, applicationId, d
         'Content-Type': 'application/json',
         'Authorization': 'Bearer ' + sessionData.connection
       }
-      appLogger.log(coSPId, 'info')
-      appLogger.log(coNetworkId, 'info')
-
       options.json = deNormalizeApplicationData(
-        applicationData.networkSettings,
+        Object.assign({}, application, applicationData.networkSettings),
         coSPId,
         coNetworkId,
         application
       )
-
       options.agentOptions = {
         'secureProtocol': 'TLSv1_2_method',
         'rejectUnauthorized': false
@@ -2161,8 +2084,7 @@ module.exports.addApplication = function (sessionData, network, applicationId, d
       request(options, async function (error, response, body) {
         if (error || response.statusCode >= 400) {
           if (error) {
-            appLogger.log('Error on create application: ', 'error')
-            appLogger.log(error, 'error')
+            appLogger.log('Error on create application: ' + error)
             reject(error)
           }
           else {
@@ -2172,7 +2094,7 @@ module.exports.addApplication = function (sessionData, network, applicationId, d
         }
         else {
           try {
-          // Save the application ID from the remote network.
+            // Save the application ID from the remote network.
             await dataAPI.putProtocolDataForKey(network.id,
               network.networkProtocol.id,
               makeApplicationDataKey(application.id, 'appNwkId'),
@@ -2226,8 +2148,7 @@ module.exports.getApplication = function (sessionData, network, applicationId, d
 
     request(options, function (error, response, body) {
       if (error) {
-        appLogger.log('Error on get application: ', 'error')
-        appLogger.log(error, 'error')
+        appLogger.log('Error on get application: ' + error)
         reject(error)
       }
       else {
@@ -2253,16 +2174,16 @@ module.exports.getApplication = function (sessionData, network, applicationId, d
 module.exports.updateApplication = function (sessionData, network, applicationId, dataAPI) {
   return new Promise(async function (resolve, reject) {
     // Get the application data.
-    var application = await dataAPI.getApplicationById(applicationId)
-    var coNetworkId = await dataAPI.getProtocolDataForKey(
+    let application = await dataAPI.getApplicationById(applicationId)
+    let coNetworkId = await dataAPI.getProtocolDataForKey(
       network.id,
       network.networkProtocol.id,
       makeCompanyDataKey(application.company.id, 'coNwkId'))
-    var appNetworkId = await dataAPI.getProtocolDataForKey(
+    let appNetworkId = await dataAPI.getProtocolDataForKey(
       network.id,
       network.networkProtocol.id,
       makeApplicationDataKey(applicationId, 'appNwkId'))
-    var applicationData = await dataAPI.getApplicationNetworkType(applicationId, network.networkType.id)
+    let applicationData = await dataAPI.getApplicationNetworkType(applicationId, network.networkType.id)
     let coSPId = await dataAPI.getProtocolDataForKey(
       network.id,
       network.networkProtocol.id,
@@ -2285,7 +2206,13 @@ module.exports.updateApplication = function (sessionData, network, applicationId
     //   'payloadDecoderScript': '',
     //   'payloadEncoderScript': ''
     // }
-    options.json = deNormalizeApplicationData(applicationData.networkSettings, coSPId, coNetworkId, application)
+    options.json = deNormalizeApplicationData(
+      applicationData.networkSettings,
+      coSPId,
+      coNetworkId,
+      application
+    )
+
     options.agentOptions = {
       'secureProtocol': 'TLSv1_2_method',
       'rejectUnauthorized': false
@@ -2332,12 +2259,11 @@ module.exports.updateApplication = function (sessionData, network, applicationId
     }
     appLogger.log(application)
     appLogger.log(applicationData)
-    appLogger.log(options, 'warn')
+    appLogger.log(options)
 
     request(options, function (error, response, body) {
       if (error) {
-        appLogger.log('Error on update application: ', 'error')
-        appLogger.log(error, 'error')
+        appLogger.log('Error on update application: ' + error)
         reject(error)
       }
       else {
@@ -2384,8 +2310,7 @@ module.exports.deleteApplication = function (sessionData, network, applicationId
 
     request(options, async function (error, response, body) {
       if (error) {
-        appLogger.log('Error on delete application: ', 'error')
-        appLogger.log(error, 'error')
+        appLogger.log('Error on delete application: ' + error)
         reject(error)
       }
       else {
@@ -2450,8 +2375,7 @@ module.exports.startApplication = function (sessionData, network, applicationId,
 
       request(options, function (error, response, body) {
         if (error) {
-          appLogger.log('Error on add application data reporting: ', 'error')
-          appLogger.log(error, 'error')
+          appLogger.log('Error on add application data reporting: ' + error)
           reject(error)
         }
         else {
@@ -2518,8 +2442,7 @@ module.exports.stopApplication = function (sessionData, network, applicationId, 
     }
     request(options, function (error, response, body) {
       if (error) {
-        appLogger.log('Error on delete application notification: ', 'error')
-        appLogger.log(error, 'error')
+        appLogger.log('Error on delete application notification: ' + error)
         reject(error)
       }
       else {
@@ -2551,7 +2474,7 @@ module.exports.passDataToApplication = function (network, applicationId, data, d
         ' for applicationId ' + applicationId +
         '. The appliction is not in a running state.  Data = ' +
         JSON.stringify(data))
-      reject(new Error('Application not running'))
+      reject('Application not running')
       return
     }
 
@@ -2631,7 +2554,6 @@ module.exports.addDeviceProfile = function (sessionData, network, deviceProfileI
         network.networkProtocol.id,
         makeCompanyDataKey(company.id, 'coSPNwkId'))
 
-      // Set up the request options.
       var options = {}
       options.method = 'POST'
       options.url = network.baseUrl + '/device-profiles'
@@ -2649,12 +2571,11 @@ module.exports.addDeviceProfile = function (sessionData, network, deviceProfileI
         'rejectUnauthorized': false
       }
 
-      appLogger.log(options, 'warn')
+      appLogger.log(options)
       request(options, async function (error, response, body) {
         if (error || response.statusCode >= 400) {
           if (error) {
-            appLogger.log('Error on create deviceProfile ' + deviceProfile.name + ':  ', 'error')
-            appLogger.log(error, 'error')
+            appLogger.log('Error on create deviceProfile ' + deviceProfile.name + ':  ' + error)
             reject(error)
           }
           else {
@@ -2664,12 +2585,13 @@ module.exports.addDeviceProfile = function (sessionData, network, deviceProfileI
         }
         else {
           // Save the deviceProfile ID from the remote network.
+          appLogger.log(body)
           await dataAPI.putProtocolDataForKey(network.id,
             network.networkProtocol.id,
             makeDeviceProfileDataKey(deviceProfile.id, 'dpNwkId'),
-            body.id)
+            body.deviceProfileID)
 
-          resolve(body.id)
+          resolve(body.deviceProfileID)
         }
       })
     }
@@ -2722,8 +2644,7 @@ module.exports.getDeviceProfile = function (sessionData, network, deviceProfileI
     request(options, function (error, response, body) {
       if (error || response.statusCode >= 400) {
         if (error) {
-          appLogger.log('Error on get deviceProfile: ', 'error')
-          appLogger.log(error, 'error')
+          appLogger.log('Error on get deviceProfile: ' + error)
           reject(error)
         }
         else {
@@ -2860,8 +2781,7 @@ module.exports.updateDeviceProfile = function (sessionData, network, deviceProfi
     request(options, function (error, response, body) {
       if (error || response.statusCode >= 400) {
         if (error) {
-          appLogger.log('Error on put deviceProfile: ', 'error')
-          appLogger.log(error, 'error')
+          appLogger.log('Error on put deviceProfile: ' + error)
           reject(error)
         }
         else {
@@ -2925,8 +2845,7 @@ module.exports.deleteDeviceProfile = function (sessionData, network, deviceProfi
     request(options, function (error, response, body) {
       if (error || response.statusCode >= 400) {
         if (error) {
-          appLogger.log('Error on delete deviceProfile: ', 'error')
-          appLogger.log(error, 'error')
+          appLogger.log('Error on delete deviceProfile: ' + error)
           reject(error)
         }
         else {
@@ -2962,7 +2881,7 @@ module.exports.deleteDeviceProfile = function (sessionData, network, deviceProfi
 // device.  The promise returns the id of the created device to be added to the
 // deviceNetworkTypeLinks record by the caller.
 module.exports.addDevice = function (sessionData, network, deviceId, dataAPI) {
-  appLogger.log('Adding Device ' + deviceId + ' to Lora V2 Network')
+  appLogger.log('Adding Device ' + deviceId + ' to Lora V1 Network')
   return new Promise(async function (resolve, reject) {
     var device
     var dntl
@@ -2974,7 +2893,7 @@ module.exports.addDevice = function (sessionData, network, deviceId, dataAPI) {
       dntl = await dataAPI.getDeviceNetworkType(deviceId, network.networkType.id)
       deviceProfile = await dataAPI.getDeviceProfileById(dntl.deviceProfile.id)
       if (!dntl.networkSettings || !dntl.networkSettings.devEUI) {
-        appLogger.log('deviceNetworkTypeLink MUST have networkSettings which MUST have devEUI', 'error')
+        appLogger.log('deviceNetworkTypeLink MUST have networkSettings which MUST have devEUI')
         reject(400)
         return
       }
@@ -2986,10 +2905,9 @@ module.exports.addDevice = function (sessionData, network, deviceId, dataAPI) {
         network.id,
         network.networkProtocol.id,
         makeDeviceProfileDataKey(dntl.deviceProfile.id, 'dpNwkId'))
-      // let loraV2Device = deNormalizeDeviceData(dntl.networkSettings, deviceProfile.networkSettings, appNwkId, dpNwkId)
-      let loraV2Device = deNormalizeDeviceData(
+
+      let loraV1Device = deNormalizeDeviceData(
         Object.assign({}, device, dntl.networkSettings),
-        deviceProfile.networkSettings,
         appNwkId,
         dpNwkId
       )
@@ -3001,65 +2919,71 @@ module.exports.addDevice = function (sessionData, network, deviceId, dataAPI) {
         'Content-Type': 'application/json',
         'Authorization': 'Bearer ' + sessionData.connection
       }
-      options.json = {
-        device: loraV2Device.device
-      }
+      // options.json = {
+      //   'applicationID': appNwkId,
+      //   'description': loraV1Device.description,
+      //   'devEUI': loraV1Device.devEUI,
+      //   'deviceProfileID': dpNwkId,
+      //   'name': loraV1Device.name,
+      //   'skipFCntCheck': loraV1Device.skipFCntCheck
+      // }
+      options.json = loraV1Device
       options.agentOptions = {
         'secureProtocol': 'TLSv1_2_method',
         'rejectUnauthorized': false
       }
 
-      appLogger.log(options, 'warn')
+      appLogger.log(options, 'info')
       request(options, function (error, response, body) {
         if (error || response.statusCode >= 400) {
           if (error) {
-            appLogger.log('Error on create device: ', 'error')
-            appLogger.lop(error, 'error')
-            appLogger.log(error, 'error')
+            appLogger.log('Error on create device: ' + error)
             reject(error)
           }
           else {
-            appLogger.log('Error on create device (' + response.statusCode + '): ', 'error')
-            appLogger.log(body, 'error')
+            appLogger.log('Error on create device (' + response.statusCode + '): ' + body.error)
             reject(response.statusCode)
           }
         }
         else {
+        // LoRa Server uses the DevEUI as the node id.
           dataAPI.putProtocolDataForKey(network.id,
             network.networkProtocol.id,
             makeDeviceDataKey(device.id, 'devNwkId'),
-            loraV2Device.device.devEUI)
+            options.json.devEUI)
 
-          appLogger.log(loraV2Device, 'warn')
+          appLogger.log(loraV1Device, 'info')
 
-          if (deviceProfile.networkSettings.supportsJoin && loraV2Device.deviceKeys) {
+          // Devices have to do a second call to set up either the
+          // Application Key (OTAA) or the Keys for ABP.
+          if (deviceProfile.networkSettings.supportsJoin && loraV1Device.deviceKeys) {
+            // This is the OTAA path.
             options.url = network.baseUrl + '/devices/' +
-              loraV2Device.device.devEUI + '/keys'
+              loraV1Device.devEUI + '/keys'
             options.json = {
-              deviceKeys: loraV2Device.deviceKeys
+              devEUI: loraV1Device.devEUI,
+              deviceKeys: loraV1Device.deviceKeys
             }
           }
-          else if (loraV2Device.deviceActivation) {
-            options.url = network.baseUrl + '/devices/' + loraV2Device.device.devEUI + '/activate'
-            options.json = {
-              deviceActivation: loraV2Device.deviceActivation
-            }
+          else if (loraV1Device.deviceActivation) {
+            // This is the ABP path.
+            options.url = network.baseUrl + '/devices/' +
+              loraV1Device.devEUI + '/activate'
+            options.json = loraV1Device.deviceActivation
+            appLogger.log('options.json = ' + JSON.stringify(options.json))
           }
           else {
-            appLogger.log('Remote Device ' + loraV2Device.name + ' does not have authentication parameters', 'error')
+            appLogger.log('Remote Device ' + loraV1Device.name + ' does not have authentication parameters')
             resolve(dntl.networkSettings.devEUI)
             return
           }
-          appLogger.log(options, 'warn')
           request(options, function (error, response, body) {
             if (error || response.statusCode >= 400) {
               if (error) {
-                appLogger.log('Error on create device keys/activation: ', 'error')
-                appLogger.log(error)
+                appLogger.log('Error on create device keys/activation: ' + error)
               }
               else {
-                appLogger.log('Error on create device keys/activation (' + response.statusCode + '): ', 'error')
-                appLogger.log(body, 'error')
+                appLogger.log('Error on create device keys/activation (' + response.statusCode + '): ' + body.error)
               }
               resolve(dntl.networkSettings.devEUI)
             }
@@ -3071,8 +2995,7 @@ module.exports.addDevice = function (sessionData, network, deviceId, dataAPI) {
       })
     }
     catch (err) {
-      appLogger.log('Error getting data for remote network: ', 'error')
-      appLogger.log(err, 'error')
+      appLogger.log('Error getting data for remote network: ' + err)
       reject(err)
     }
   })
@@ -3120,8 +3043,7 @@ module.exports.getDevice = function (sessionData, network, deviceId, dataAPI) {
     request(options, function (error, response, body) {
       if (error || response.statusCode >= 400) {
         if (error) {
-          appLogger.log('Error on get device: ', 'error')
-          appLogger.log(error, 'error')
+          appLogger.log('Error on get device: ' + error)
           reject(error)
         }
         else {
@@ -3191,13 +3113,12 @@ module.exports.updateDevice = function (sessionData, network, deviceId, dataAPI)
       //   'deviceProfileID': dpNwkId,
       //   'name': device.name
       // }
-      let loraV2Device = deNormalizeDeviceData(
+      let loraV1Device = deNormalizeDeviceData(
         Object.assign({}, device, dntl.networkSettings),
-        deviceProfile,
         appNwkId,
         dpNwkId
       )
-      options.json = loraV2Device
+      options.json = loraV1Device
       options.agentOptions = {
         'secureProtocol': 'TLSv1_2_method',
         'rejectUnauthorized': false
@@ -3219,24 +3140,24 @@ module.exports.updateDevice = function (sessionData, network, deviceId, dataAPI)
         else {
           // Devices have to do a second call to set up either the
           // Application Key (OTAA) or the Keys for ABP.
-          if (deviceProfile.networkSettings.supportsJoin && loraV2Device.deviceKeys) {
+          if (deviceProfile.networkSettings.supportsJoin && loraV1Device.deviceKeys) {
             // This is the OTAA path.
             options.url = network.baseUrl + '/devices/' +
-              loraV2Device.devEUI + '/keys'
+              loraV1Device.devEUI + '/keys'
             options.json = {
-              devEUI: loraV2Device.devEUI,
-              deviceKeys: loraV2Device.deviceKeys
+              devEUI: loraV1Device.devEUI,
+              deviceKeys: loraV1Device.deviceKeys
             }
           }
-          else if (loraV2Device.deviceActivation) {
+          else if (loraV1Device.deviceActivation) {
             // This is the ABP path.
             options.url = network.baseUrl + '/devices/' +
-              loraV2Device.devEUI + '/activate'
-            options.json = loraV2Device.deviceActivation
+              loraV1Device.devEUI + '/activate'
+            options.json = loraV1Device.deviceActivation
             appLogger.log('options.json = ' + JSON.stringify(options.json))
           }
           else {
-            appLogger.log('Remote Device ' + loraV2Device.name + ' does not have authentication parameters')
+            appLogger.log('Remote Device ' + loraV1Device.name + ' does not have authentication parameters')
             resolve(dntl.networkSettings.devEUI)
             return
           }
@@ -3310,8 +3231,7 @@ module.exports.deleteDevice = function (sessionData, network, deviceId, dataAPI)
     request(options, async function (error, response, body) {
       if (error || response.statusCode >= 400) {
         if (error) {
-          appLogger.log('Error on delete device: ', 'error')
-          appLogger.log(error, 'error')
+          appLogger.log('Error on delete device: ' + error)
           reject(error)
         }
         else {
@@ -3336,7 +3256,7 @@ module.exports.deleteDevice = function (sessionData, network, deviceId, dataAPI)
         request(options, function (error, response, body) {
           if (error || response.statusCode >= 400) {
             if (error) {
-              appLogger.log('Error on delete device keys: ', 'error')
+              appLogger.log('Error on delete device keys: ' + error)
             }
             else {
               appLogger.log('Error on delete device keys (' + response.statusCode + '): ' + body.error)
@@ -3355,26 +3275,24 @@ module.exports.deleteDevice = function (sessionData, network, deviceId, dataAPI)
 
 function normalizeApplicationData (remoteApplication) {
   /*
- "application": {
-        "description": "string",
-        "id": "string",
-        "name": "string",
-        "organizationID": "string",
-        "payloadCodec": "string",
-        "payloadDecoderScript": "string",
-        "payloadEncoderScript": "string",
-        "serviceProfileID": "string"
-      }
+  "description": "string",
+      "id": "string",
+      "name": "string",
+      "organizationID": "string",
+      "payloadCodec": "string",
+      "payloadDecoderScript": "string",
+      "payloadEncoderScript": "string",
+      "serviceProfileID": "string"
    */
   let normalized = {
-    description: remoteApplication.application.description,
-    id: remoteApplication.application.id,
-    name: remoteApplication.application.name,
-    organizationID: remoteApplication.application.organizationID,
-    payloadCodec: remoteApplication.application.payloadCodec,
-    payloadDecoderScript: remoteApplication.application.payloadDecoderScript,
-    payloadEncoderScript: remoteApplication.application.payloadEncoderScript,
-    serviceProfileID: remoteApplication.application.serviceProfileID,
+    description: remoteApplication.description,
+    id: remoteApplication.id,
+    name: remoteApplication.name,
+    organizationID: remoteApplication.organizationID,
+    payloadCodec: remoteApplication.payloadCodec,
+    payloadDecoderScript: remoteApplication.payloadDecoderScript,
+    payloadEncoderScript: remoteApplication.payloadEncoderScript,
+    serviceProfileID: remoteApplication.serviceProfileID,
     cansend: true,
     deviceLimit: null,
     devices: null,
@@ -3390,63 +3308,64 @@ function normalizeApplicationData (remoteApplication) {
 }
 
 function deNormalizeApplicationData (remoteApplication, serviceProfile, organizationId, application) {
-  let loraV2ApplicationData = {
-    application: {
-      'description': application.description,
-      'name': application.name,
-      'organizationID': organizationId,
-      'payloadCodec': remoteApplication.payloadCodec,
-      'payloadDecoderScript': remoteApplication.payloadDecoderScript,
-      'payloadEncoderScript': remoteApplication.payloadEncoderScript,
-      'serviceProfileID': serviceProfile
-    }
+  let loraV1ApplicationData = {
+    'description': application.description || remoteApplication.description,
+    'id': remoteApplication.id,
+    'name': application.name || remoteApplication.name,
+    'organizationID': organizationId,
+    'payloadCodec': remoteApplication.payloadCodec,
+    'payloadDecoderScript': remoteApplication.payloadDecoderScript,
+    'payloadEncoderScript': remoteApplication.payloadEncoderScript,
+    'serviceProfileID': serviceProfile
   }
-  return loraV2ApplicationData
+  appLogger.log(remoteApplication, 'info')
+  appLogger.log(loraV1ApplicationData, 'info')
+  return loraV1ApplicationData
 }
 
 function normalizeDeviceProfileData (remoteDeviceProfile) {
   /*
-  "createdAt": "2018-09-05T05:28:09.681Z",
-      "deviceProfile": {
-        "classBTimeout": 0,
-        "classCTimeout": 0,
-        "factoryPresetFreqs": [
-          0
-        ],
-        "id": "string",
-        "macVersion": "string",
-        "maxDutyCycle": 0,
-        "maxEIRP": 0,
-        "name": "string",
-        "networkServerID": "string",
-        "organizationID": "string",
-        "pingSlotDR": 0,
-        "pingSlotFreq": 0,
-        "pingSlotPeriod": 0,
-        "regParamsRevision": "string",
-        "rfRegion": "string",
-        "rxDROffset1": 0,
-        "rxDataRate2": 0,
-        "rxDelay1": 0,
-        "rxFreq2": 0,
-        "supports32BitFCnt": true,
-        "supportsClassB": true,
-        "supportsClassC": true,
-        "supportsJoin": true
-      },
-      "updatedAt": "2018-09-05T05:28:09.682Z"
+  "createdAt": "string",
+    "deviceProfile": {
+      "classBTimeout": 0,
+      "classCTimeout": 0,
+      "deviceProfileID": "string",
+      "factoryPresetFreqs": [
+        0
+      ],
+      "macVersion": "string",
+      "maxDutyCycle": 0,
+      "maxEIRP": 0,
+      "pingSlotDR": 0,
+      "pingSlotFreq": 0,
+      "pingSlotPeriod": 0,
+      "regParamsRevision": "string",
+      "rfRegion": "string",
+      "rxDROffset1": 0,
+      "rxDataRate2": 0,
+      "rxDelay1": 0,
+      "rxFreq2": 0,
+      "supports32bitFCnt": true,
+      "supportsClassB": true,
+      "supportsClassC": true,
+      "supportsJoin": true
+    },
+    "name": "string",
+    "networkServerID": "string",
+    "organizationID": "string",
+    "updatedAt": "string"
    */
   let normalized = {
     classBTimeout: remoteDeviceProfile.deviceProfile.classBTimeout,
     classCTimeout: remoteDeviceProfile.deviceProfile.classCTimeout,
     factoryPresetFreqs: remoteDeviceProfile.deviceProfile.factoryPresetFreqs,
-    id: remoteDeviceProfile.deviceProfile.id,
+    id: remoteDeviceProfile.deviceProfile.deviceProfileID,
     macVersion: remoteDeviceProfile.deviceProfile.macVersion,
     maxDutyCycle: remoteDeviceProfile.deviceProfile.maxDutyCycle,
     maxEIRP: remoteDeviceProfile.deviceProfile.maxEIRP,
-    name: remoteDeviceProfile.deviceProfile.name,
-    networkServerID: remoteDeviceProfile.deviceProfile.networkServerID,
-    organizationID: remoteDeviceProfile.deviceProfile.organizationID,
+    name: remoteDeviceProfile.name,
+    networkServerID: remoteDeviceProfile.networkServerID,
+    organizationID: remoteDeviceProfile.organizationID,
     pingSlotDR: remoteDeviceProfile.deviceProfile.pingSlotDR,
     pingSlotFreq: remoteDeviceProfile.deviceProfile.pingSlotFreq,
     pingSlotPeriod: remoteDeviceProfile.deviceProfile.pingSlotPeriod,
@@ -3465,41 +3384,11 @@ function normalizeDeviceProfileData (remoteDeviceProfile) {
 }
 
 function deNormalizeDeviceProfileData (remoteDeviceProfile, networkServerId, organizationId) {
-  /*
-    "createdAt": "2018-09-05T05:28:09.681Z",
-      "deviceProfile": {
-        "classBTimeout": 0,
-        "classCTimeout": 0,
-        "factoryPresetFreqs": [
-          0
-        ],
-        "id": "string",
-        "macVersion": "string",
-        "maxDutyCycle": 0,
-        "maxEIRP": 0,
-        "name": "string",
-        "networkServerID": "string",
-        "organizationID": "string",
-        "pingSlotDR": 0,
-        "pingSlotFreq": 0,
-        "pingSlotPeriod": 0,
-        "regParamsRevision": "string",
-        "rfRegion": "string",
-        "rxDROffset1": 0,
-        "rxDataRate2": 0,
-        "rxDelay1": 0,
-        "rxFreq2": 0,
-        "supports32BitFCnt": true,
-        "supportsClassB": true,
-        "supportsClassC": true,
-        "supportsJoin": true
-      },
-      "updatedAt": "2018-09-05T05:28:09.682Z"
-   */
-  let loraV2DeviceProfileData = {
+  let loraV1DeviceProfileData = {
     deviceProfile: {
       classBTimeout: remoteDeviceProfile.classBTimeout,
       classCTimeout: remoteDeviceProfile.classCTimeout,
+      deviceProfileID: `${remoteDeviceProfile.id}`,
       factoryPresetFreqs: remoteDeviceProfile.factoryPresetFreqs,
       macVersion: remoteDeviceProfile.macVersion,
       maxDutyCycle: remoteDeviceProfile.maxDutyCycle,
@@ -3513,129 +3402,111 @@ function deNormalizeDeviceProfileData (remoteDeviceProfile, networkServerId, org
       rxDataRate2: remoteDeviceProfile.rxDataRate2,
       rxDelay1: remoteDeviceProfile.rxDelay1,
       rxFreq2: remoteDeviceProfile.rxFreq2,
-      supports32BitFCnt: remoteDeviceProfile.supports32BitFCnt,
+      supports32bitFCnt: remoteDeviceProfile.supports32BitFCnt,
       supportsClassB: remoteDeviceProfile.supportsClassB,
       supportsClassC: remoteDeviceProfile.supportsClassC,
-      supportsJoin: remoteDeviceProfile.supportsJoin,
-      name: remoteDeviceProfile.name,
-      description: remoteDeviceProfile.description,
-      networkServerID: networkServerId,
-      organizationID: organizationId
-    }
-
+      supportsJoin: remoteDeviceProfile.supportsJoin
+    },
+    name: remoteDeviceProfile.name,
+    networkServerID: networkServerId,
+    organizationID: organizationId
   }
-  return loraV2DeviceProfileData
+  return loraV1DeviceProfileData
 }
 
 function normalizeDeviceData (remoteDevice) {
   /*
-        "applicationID": "string",
+      "applicationID": "string",
       "description": "string",
       "devEUI": "string",
       "deviceProfileID": "string",
-      "name": "string",
-      "skipFCntCheck": true,
       "deviceStatusBattery": 0,
       "deviceStatusMargin": 0,
-      "lastSeenAt": "2018-09-05T05:28:09.738Z"
-"deviceActivation": {
-    "aFCntDown": 0,
-    "appSKey": "string",
-    "devAddr": "string",
-    "devEUI": "string",
-    "fCntUp": 0,
-    "fNwkSIntKey": "string",
-    "nFCntDown": 0,
-    "nwkSEncKey": "string",
-    "sNwkSIntKey": "string"
-  }
-   "deviceKeys": {
-    "appKey": "string",
-    "devEUI": "string",
-    "nwkKey": "string"
-  }
+      "lastSeenAt": "string",
+      "name": "string",
+      "skipFCntCheck": true
    */
   let normalized = {
-    applicationID: remoteDevice.device.applicationID,
-    description: remoteDevice.device.description,
-    devEUI: remoteDevice.device.devEUI,
-    deviceProfileID: remoteDevice.device.deviceProfileID,
-    name: remoteDevice.device.name,
-    skipFCntCheck: remoteDevice.device.skipFCntCheck,
+    applicationID: remoteDevice.applicationID,
+    description: remoteDevice.description,
+    devEUI: remoteDevice.devEUI,
+    deviceProfileID: remoteDevice.deviceProfileID,
+    name: remoteDevice.name,
+    skipFCntCheck: remoteDevice.skipFCntCheck,
     deviceStatusBattery: remoteDevice.deviceStatusBattery,
     deviceStatusMargin: remoteDevice.deviceStatusMargin,
     lastSeenAt: remoteDevice.lastSeenAt
   }
-  if (remoteDevice.device.deviceKeys) {
+  if (remoteDevice.deviceKeys) {
     normalized.deviceKeys = {
-      appKey: remoteDevice.device.deviceKeys.appKey,
-      devEUI: remoteDevice.device.deviceKeys.devEUI,
-      nwkKey: remoteDevice.device.deviceKeys.nwkKey
+      appKey: remoteDevice.deviceKeys.appKey,
+      devEUI: remoteDevice.deviceKeys.devEUI,
+      nwkKey: remoteDevice.deviceKeys.appKey
     }
   }
-
-  if (remoteDevice.device.deviceActivation) {
+  /*
+  "devEUI": "",
+  "appSKey": "string",
+  "devAddr": "string",
+  "fCntDown": 0,
+  "fCntUp": 0,
+  "nwkSKey": "string",
+  "skipFCntCheck": true
+   */
+  if (remoteDevice.deviceActivation) {
     normalized.deviceActivation = {
-      aFCntDown: remoteDevice.device.deviceActivation.aFCntDown,
-      appSKey: remoteDevice.device.deviceActivation.appSKey,
-      devAddr: remoteDevice.device.deviceActivation.devAddr,
-      devEUI: remoteDevice.device.deviceActivation.devEUI,
-      fCntUp: remoteDevice.device.deviceActivation.fCntUp,
-      nFCntDown: remoteDevice.device.deviceActivation.nFCntDown,
-      nwkSEncKey: remoteDevice.device.deviceActivation.nwkSEncKey,
-      sNwkSIntKey: remoteDevice.device.deviceActivation.sNwkSIntKey,
-      fNwkSIntKey: remoteDevice.device.deviceActivation.fNwkSIntKey
+      devEUI: remoteDevice.deviceActivation.devEUI,
+      appSKey: remoteDevice.deviceActivation.appSKey,
+      devAddr: remoteDevice.deviceActivation.devAddr,
+      aFCntDown: remoteDevice.deviceActivation.fCntDown,
+      nFCntDown: remoteDevice.deviceActivation.fCntDown,
+      fCntUp: remoteDevice.deviceActivation.fCntUp,
+      nwkSEncKey: remoteDevice.deviceActivation.nwkSKey,
+      sNwkSIntKey: remoteDevice.deviceActivation.nwkSKey,
+      fNwkSIntKey: remoteDevice.deviceActivation.nwkSKey
     }
   }
   return normalized
 }
 
-/**
- * @see ./data/json/lora2.json
- *
- * @param remoteDevice
- * @returns {{device: {applicationID: (*|string), description: *, devEUI: *, deviceProfileID: *, name: *, skipFCntCheck: (*|boolean)}, deviceStatusBattery: number, deviceStatusMargin: number, lastSeenAt: (string|null)}}
- */
-function deNormalizeDeviceData (remoteDevice, deviceProfile, appId, dpId) {
-  let loraV2DeviceData = {
-    device: {
-      applicationID: appId,
-      description: remoteDevice.description,
-      devEUI: remoteDevice.devEUI,
-      deviceProfileID: dpId,
-      name: remoteDevice.name,
-      skipFCntCheck: remoteDevice.skipFCntCheck
-    }
+function deNormalizeDeviceData (remoteDevice, appId, dpId) {
+  /*
+      "applicationID": "string",
+      "description": "string",
+      "devEUI": "string",
+      "deviceProfileID": "string",
+      "deviceStatusBattery": 0,
+      "deviceStatusMargin": 0,
+      "lastSeenAt": "string",
+      "name": "string",
+      "skipFCntCheck": true
+   */
+  let loraV1DeviceData = {
+    applicationID: appId,
+    description: remoteDevice.description,
+    devEUI: remoteDevice.devEUI,
+    deviceProfileID: dpId,
+    name: remoteDevice.name,
+    skipFCntCheck: remoteDevice.skipFCntCheck,
+    deviceStatusBattery: remoteDevice.deviceStatusBattery,
+    deviceStatusMargin: remoteDevice.deviceStatusMargin,
+    lastSeenAt: remoteDevice.lastSeenAt
   }
   if (remoteDevice.deviceKeys) {
-    if (deviceProfile.macVersion === '1.1.0') {
-      loraV2DeviceData.deviceKeys = {
-        appKey: remoteDevice.deviceKeys.appKey,
-        nwkKey: remoteDevice.deviceKeys.nwkKey,
-        devEUI: remoteDevice.deviceKeys.devEUI
-      }
-    }
-    else {
-      loraV2DeviceData.deviceKeys = {
-        nwkKey: remoteDevice.deviceKeys.appKey,
-        devEUI: remoteDevice.deviceKeys.devEUI,
-        appKey: remoteDevice.deviceKeys.appKey
-      }
+    loraV1DeviceData.deviceKeys = {
+      appKey: remoteDevice.deviceKeys.appKey
     }
   }
   if (remoteDevice.deviceActivation) {
-    loraV2DeviceData.deviceActivation = {
+    loraV1DeviceData.deviceActivation = {
       appSKey: remoteDevice.deviceActivation.appSKey,
       devAddr: remoteDevice.deviceActivation.devAddr,
-      aFCntDown: remoteDevice.deviceActivation.aFCntDown,
-      nFCntDown: remoteDevice.deviceActivation.nFCntDown,
+      devEUI: remoteDevice.deviceActivation.devEUI,
+      fCntDown: remoteDevice.deviceActivation.aFCntDown,
       fCntUp: remoteDevice.deviceActivation.fCntUp,
-      nwkSEncKey: remoteDevice.deviceActivation.nwkSEncKey,
-      sNwkSIntKey: remoteDevice.deviceActivation.sNwkSIntKey,
-      fNwkSIntKey: remoteDevice.deviceActivation.fNwkSIntKey
+      nwkSKey: remoteDevice.deviceActivation.nwkSEncKey,
+      skipFCntCheck: false
     }
   }
-  appLogger.log(remoteDevice, 'warn')
-  appLogger.log(loraV2DeviceData, 'warn')
-  return loraV2DeviceData
+  return loraV1DeviceData
 }
