+++ rest/networkProtocols/handlers/README.md
@@ -1,0 +1,4 @@
+# Network Protocol Handlers
+
+Handlers are registered alphabetically.  If one protocol use's another
+as it's master protocol, the master protocol should be listed first.
