+++ rest/networkProtocols/handlers/TheThingsNetwork.js
@@ -1,6 +1,6 @@
 const request = require('request-promise')
 const nconf = require('nconf')
-const appLogger = require('../lib/appLogger.js')
+const appLogger = require('../../lib/appLogger.js')
 const uuid = require('uuid/v1')
 const R = require('ramda')
