+++ test2.0/api/001-restSessions.js
@@ -7,67 +7,53 @@ var should = chai.should()
 chai.use(chaiHttp)
 var server = chai.request(app).keepOpen()
 
-describe('Sessions', function () {
+describe.skip('Sessions', () => {
   var adminToken
   before((done) => {
     setTimeout(done, 2000)
   })
 
-  describe('POST /api/sessions', function () {
-    it('should return 401 when the login is not valid', function (done) {
-      server
+  describe('POST /api/sessions', () => {
+    it('should return 401 when the login is not valid', async () => {
+      const res = await server
         .post('/api/sessions')
         .send({ 'login_username': 'foo', 'login_password': 'bar' })
-        .end(function (err, res) {
-          res.should.have.status(401)
-          done()
-        })
+      res.should.have.status(401)
     })
 
-    it('should return 200 and a jwt when the login is valid', function (done) {
-      server
+    it('should return 200 and a jwt when the login is valid', async () => {
+      const res = await server
         .post('/api/sessions')
         .send({ 'login_username': 'admin', 'login_password': 'password' })
-        .end(function (err, res) {
-          res.should.have.status(200)
-          adminToken = res.text
-          done()
-        })
+      res.should.have.status(200)
+      adminToken = res.text
+      console.log("ADMIN_TOKEN", adminToken)
     })
 
-    it('should be able to use the admin jwt to get all companies', function (done) {
-      server
+    it('should be able to use the admin jwt to get all companies', async () => {
+      const res = await server
         .get('/api/companies')
         .set('Authorization', 'Bearer ' + adminToken)
         .set('Content-Type', 'application/json')
-        .end(function (err, res) {
-          res.should.have.status(200)
-          var result = JSON.parse(res.text)
-          result.records.should.be.instanceof(Array)
-          done()
-        })
+      res.should.have.status(200)
+      var result = JSON.parse(res.text)
+      result.records.should.be.instanceof(Array)
     })
   })
 
-  describe('DELETE /api/sessions', function () {
-    it('should return 401 with no token', function (done) {
-      server
+  describe('DELETE /api/sessions', () => {
+    it('should return 401 with no token', async () => {
+      const res = await server
         .delete('/api/sessions')
-        .end(function (err, res) {
-          res.should.have.status(401)
-          done()
-        })
+      res.should.have.status(401)
     })
 
-    it('should return 204 when the login is valid', function (done) {
-      server
+    it('should return 204 when the login is valid', async () => {
+      const res = await server
         .delete('/api/sessions')
         .set('Authorization', 'Bearer ' + adminToken)
-        .end(function (err, res) {
-          res.should.have.status(204)
-          adminToken = {}
-          done()
-        })
+      res.should.have.status(204)
+      adminToken = {}
     })
   })
 })
